+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Innmind\Validation\Failure::tag()`
 - `Innmind\Validation\Failure::tags()`
+- `Innmind\Validation\Constraint::mapFailures()`
 
 ## 2.2.0 - 2025-09-04